refactor(types): group the Go-paired wire contracts under src/types/wire/
Moves wire-enhanced-action.ts, wire-network.ts, wire-performance-snapshot.ts
and wire-websocket-event.ts into src/types/wire/.

src/types has almost no usable co-change signal — 19 of its 22 files were
introduced by a single release squash (93d8ca74 / b47054f7), and every module
is reached through the two-level barrel index.ts -> messages.ts, so
"co-changes with" degenerates to "co-changes with everything". The reference
graph is the evidence that does discriminate, and for these four it is
unambiguous: they have zero importers anywhere outside src/types. Each is
consumed by exactly one domain wrapper that re-exports it under a friendlier
alias (actions.ts -> EnhancedAction, network.ts -> NetworkBodyPayload and
WaterfallEntry, performance.ts -> PerformanceSnapshot, websocket.ts ->
WebSocketEvent) plus the index barrel. Nothing else in the repo imports a
Wire* symbol directly. They are a closed set.

They are also the only files in src/ that are generated from Go rather than
written: scripts/generate-wire-types.js emits them from internal/types/
wire_*.go and internal/performance/wire_performance.go, and
scripts/check-wire-drift.js gates them (rule 8). Giving that contract its own
folder makes the boundary visible instead of leaving it encoded in a filename
prefix.

Behavior is unchanged: `node scripts/check-wire-drift.js` reports "OK: 13 wire
type pairs verified, zero drift" after the move, and the regenerated files are
byte-identical apart from location.

Non-obvious path references updated — both wire scripts hold the TS path as a
data literal, so nothing about the move is visible to TypeScript:
  - scripts/generate-wire-types.js PAIRS (4 entries)
  - scripts/check-wire-drift.js pair table (4 entries)
  - docs: the idl-migration feature index now lists the four TS files in
    code_paths, plus design-spec.md and the four architecture notes that name
    src/types/wire-*.ts (server-architecture, extension-architecture,
    invariants ARCH_INV_005, interfaces)

scripts/check-sync-wire-drift.js was checked and needs no change: it compares
internal/capture/sync.go against src/background/sync/sync-client.ts and never
mentions src/types.

16 stale compiled artifacts under extension/types/ were git rm'd; the rebuilt
output lands in extension/types/wire/.
