+++ docs/architecture/invariants.md
@@ -18,7 +18,7 @@ These are non-negotiable system behaviors. Any change that violates one requires
 | `ARCH_INV_002` | Observe/interact command-status handlers keep real queue-backed behavior and never regress to stubs. | `cmd/browser-agent/tools_observe.go`, `cmd/browser-agent/tools_interact.go`, `cmd/browser-agent/tools_core.go` | `./scripts/validate-architecture.sh`, `go test ./cmd/browser-agent/...` |
 | `ARCH_INV_003` | MCP request/response envelope remains valid JSON-RPC 2.0. | `cmd/browser-agent/handler.go`, `internal/mcp/` | `go test ./cmd/browser-agent/...` |
 | `ARCH_INV_004` | Public MCP tool schemas remain backward compatible unless there is an explicit, versioned breaking-change plan. | `cmd/browser-agent/tools_schema.go`, `cmd/browser-agent/tools_*_schema.go` | `go test ./cmd/browser-agent/...`, docs update in `interfaces.md` |
-| `ARCH_INV_005` | Go wire types and TypeScript wire types stay in sync. | `internal/types/wire_*.go`, `src/types/wire-*.ts` | `make check-wire-drift` |
+| `ARCH_INV_005` | Go wire types and TypeScript wire types stay in sync. | `internal/types/wire_*.go`, `src/types/wire/wire-*.ts` | `make check-wire-drift` |
 | `ARCH_INV_006` | Extension/server sync protocol changes are coordinated in one change set. | `internal/capture/sync.go`, `internal/capture/handlers.go`, `src/background/sync/sync-client.ts` | `go test ./internal/capture/...` and extension test target |
 | `ARCH_INV_007` | Sensitive data handling preserves redaction guarantees in logs and surfaced outputs. | `internal/redaction/`, `cmd/browser-agent/tools_observe.go`, export paths | `go test ./internal/redaction/...` plus relevant tool tests |
 | `ARCH_INV_008` | Bridge, daemon, and CLI modes remain operable; bridge respawn behavior is preserved. | `cmd/browser-agent/bridge.go`, `cmd/browser-agent/main_connection_mcp.go`, `cmd/browser-agent/cli.go` | `go test ./cmd/browser-agent/...` |
