+++ extension/types/network.d.ts
@@ -21,11 +21,11 @@ export interface PendingRequest {
  * The stale interface previously used camelCase fields (contentType, requestBody, etc.)
  * that didn't match the Go server expectations.
  */
-export type { WireNetworkBody as NetworkBodyPayload } from './wire-network.js';
+export type { WireNetworkBody as NetworkBodyPayload } from './wire/wire-network.js';
 /**
  * Network waterfall entry — re-exported from wire type.
  * The stale interface previously used camelCase fields and a WaterfallPhases sub-object
  * that didn't match the actual runtime data.
  */
-export type { WireNetworkWaterfallEntry as WaterfallEntry } from './wire-network.js';
+export type { WireNetworkWaterfallEntry as WaterfallEntry } from './wire/wire-network.js';
 //# sourceMappingURL=network.d.ts.map
