+++ extension/types/wire/wire-network.d.ts
