+++ extension/types/wire/wire-websocket-event.js
