+++ src/types/index.ts
@@ -144,11 +144,11 @@ export type {
 } from './messages.js'
 
 // Re-export wire types (canonical HTTP payload shapes)
-export type { WireEnhancedAction } from './wire-enhanced-action.js'
+export type { WireEnhancedAction } from './wire/wire-enhanced-action.js'
 
-export type { WireNetworkBody, WireNetworkWaterfallEntry, WireNetworkWaterfallPayload } from './wire-network.js'
+export type { WireNetworkBody, WireNetworkWaterfallEntry, WireNetworkWaterfallPayload } from './wire/wire-network.js'
 
-export type { WireWebSocketEvent } from './wire-websocket-event.js'
+export type { WireWebSocketEvent } from './wire/wire-websocket-event.js'
 
 export type {
   WirePerformanceTiming,
@@ -159,7 +159,7 @@ export type {
   WireUserTimingEntry,
   WireUserTimingData,
   WirePerformanceSnapshot
-} from './wire-performance-snapshot.js'
+} from './wire/wire-performance-snapshot.js'
 
 // Re-export all utility types
 export type {
