+++ src/types/wire/wire-network.ts
