+++ src/types/wire/wire-websocket-event.ts
